Create cli interface
