+++ cv-to-pdf/.gitignore
@@ -2,3 +2,4 @@
 *.html
 *.lock
 *.json
+**/__pycache__/
